api/apiserver: implement juju-run actions
