+++ api/action/run.go
@@ -1,0 +1,27 @@
+// Copyright 2014 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package action
+
+import (
+	"time"
+
+	"github.com/juju/juju/apiserver/params"
+)
+
+// RunOnAllMachines runs the command on all the machines with the specified
+// timeout.
+func (c *Client) RunOnAllMachines(commands string, timeout time.Duration) ([]params.ActionResult, error) {
+	var results params.ActionResults
+	args := params.RunParams{Commands: commands, Timeout: timeout}
+	err := c.facade.FacadeCall("RunOnAllMachines", args, &results)
+	return results.Results, err
+}
+
+// Run the Commands specified on the machines identified through the ids
+// provided in the machines, services and units slices.
+func (c *Client) Run(run params.RunParams) ([]params.ActionResult, error) {
+	var results params.ActionResults
+	err := c.facade.FacadeCall("Run", run, &results)
+	return results.Results, err
+}
