+++ apiserver/action/export_test.go
@@ -1,0 +1,9 @@
+// Copyright 2012, 2013 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package action
+
+var (
+	GetAllUnitNames = getAllUnitNames
+	QueueActions    = &queueActions
+)
